+++ R/codebook.R
@@ -387,53 +387,54 @@ codebook_table <- function(results) {
 }
 
 metadata <- function(results) {
-  dplyr::bind_rows(
-    # var = results$session
-    lapply(results, function(var) {
-      x <- attributes(var)
-      if (is.null(x)) {
-        data.frame(label = NA)
-      } else {
-        if (exists("class", x)) {
-          x$class <- NULL
-        }
-        if (exists("tzone", x)) {
-          x$tzone <- NULL
-        }
-        if (exists("label", x)) {
-          if (exists("item", x)) {
-            if (exists("label", x$item)) {
-              x$item$label <- NULL
-            }
-            if (exists("label_parsed", x$item)) {
-              x$item$label_parsed <- NULL
-            }
-          }
-        }
-        if (exists("labels", x)) {
-          if (!is.null(names(x$labels))) {
-            x$value_labels <- paste(paste0(names(x$labels),
-                                           "=", x$labels), collapse = ",\n")
-          } else {
-            x$value_labels <- paste(x$labels, collapse = ",\n")
-          }
-          x$labels <- NULL
-          if (exists("item", x) && exists("choices", x$item)) {
-            x$item$choices <- NULL
-          }
-        }
-
-        if (exists("item", x) && exists("name", x$item)) {
-          x$item$name <- NULL
-        }
-        if (exists("scale_item_names", x)) {
-          x$scale_item_names <- paste(x$scale_item_names, collapse = ", ")
-        }
-        as.data.frame(t(purrr::flatten(x)))
-      }
-    }), .id = "name")
+    purrr::map_dfr(results, attribute_summary, .id = "name")
 }
 
+attribute_summary <- function(var) {
+  x <- attributes(var)
+  if (is.null(x)) {
+    return(data.frame(label = NA_character_))
+  }
+  if (exists("class", x)) {
+    x$class <- NULL
+  }
+  if (exists("tzone", x)) {
+    x$tzone <- NULL
+  }
+  if (exists("label", x)) {
+    if (exists("item", x)) {
+      if (exists("label", x$item)) {
+        x$item$label <- NULL
+      }
+      if (exists("label_parsed", x$item)) {
+        x$item$label_parsed <- NULL
+      }
+    }
+  }
+  if (exists("labels", x)) {
+    if (!is.null(names(x$labels))) {
+      x$value_labels <- paste(paste0(names(x$labels),
+                                     "=", x$labels), collapse = ",\n")
+    } else {
+      x$value_labels <- paste(x$labels, collapse = ",\n")
+    }
+    x$labels <- NULL
+    if (exists("item", x) && exists("choices", x$item)) {
+      x$item$choices <- NULL
+    }
+  }
+
+  if (exists("item", x) && exists("name", x$item)) {
+    x$item$name <- NULL
+  }
+  if (exists("scale_item_names", x)) {
+    x$scale_item_names <- paste(x$scale_item_names, collapse = ", ")
+  }
+  x <- purrr::flatten_dfr(purrr::flatten(x))
+  if (ncol(x) == 0) {
+    x <- data.frame(label = NA_character_)
+  }
+}
 
 
 #' Metadata from dataframe