+++ tests/testthat/codebook_browser/app.R
@@ -1,0 +1,4 @@
+library(codebook)
+
+data("bfi")
+codebook_browser(data = bfi, viewer = shiny::paneViewer())