+++ tests/testthat/codebook_browser/tests/mytest.R
@@ -1,0 +1,6 @@
+app <- ShinyDriver$new("../")
+app$snapshotInit("mytest")
+
+app$snapshot()
+app$setInputs(done = "click")
+app$snapshot()