+++ tests/testthat/label_browser/app.R
@@ -1,0 +1,4 @@
+library(codebook)
+
+data("bfi")
+label_browser(data = bfi, viewer = shiny::paneViewer())