+++ tests/testthat/test_addin.R
@@ -1,0 +1,18 @@
+# library(shinytest)
+#
+# context("Test addin shinyapps")
+#
+# # recordTest("tests/testthat/codebook_browser/")
+#
+# test_that("codebook browser", {
+#   skip_on_cran()
+#   expect_pass(testApp("codebook_browser"))
+# })
+#
+# # recordTest("tests/testthat/label_browser/")
+#
+# test_that("codebook browser", {
+#   skip_on_cran()
+#   expect_pass(testApp("label_browser"))
+# })
+#