+++ tests/testthat/test_codebook.R
@@ -256,6 +256,7 @@ test_that("codebook generation via helper fun", {
 codebook(codebook_data)
 ```
 "))
+
 })
 
 
@@ -287,9 +288,8 @@ test_that("codebook table generation, no attributes", {
   bfi$age <- 1:nrow(bfi)
   expect_silent(cb <- codebook_table(bfi))
   expect_identical(names(cb),
-                   c("name", "data_type", "ordered", "value_labels",
-                     "missing", "complete", "n",  "empty", "n_unique",
-                     "top_counts", "count", "median", "min", "max", "mean", "sd",
+                   c("name", "data_type", "missing", "complete", "n", "empty",
+                     "n_unique","count", "median", "min", "max", "mean", "sd",
                      "p0", "p25", "p50", "p75", "p100", "hist"))
   expect_equal(nrow(cb), ncol(bfi))
   expect_identical(cb$name, names(bfi))